+++ package.json
@@ -8,6 +8,7 @@
   },
   "dependencies": {
     "@githubnext/blocks": "^2.3.3",
+    "@githubnext/blocks-runtime": "^1.0.0",
     "@primer/react": "^35.15.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
