--- src/index.d.ts
@@ -1,4 +1,0 @@
-type Asset = {
-  name: string;
-  content: string;
-};
